update user table

+++ lara-blogs/app/Http/Controllers/PostController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\UpdatePostRequest;
 use App\Models\Category;
 use App\Models\Post;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
@@ -79,6 +80,10 @@ class PostController extends Controller
      */
     public function update(UpdatePostRequest $request, Post $post)
     {
+
+        if (Gate::denies('update', $post)) {
+            return abort(403, 'U are not allowed to update!');
+        };
         $post->title = $request->title;
         $post->slug = Str::slug($request->title);
         $post->description = $request->description;
@@ -104,6 +109,12 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+
+
+        if (Gate::denies('delete', $post)) {
+
+            return abort(403, 'U are not allowed to delete');
+        };
         $postTitle = $post->title;
         Storage::delete($post->featured_image);
         $post->delete();
